test(test/*): update test script
