+++ test/index.test.ts
@@ -2,6 +2,40 @@ import { Application, Framework } from '@midwayjs/koa';
 import { close, createApp, createHttpRequest } from '@midwayjs/mock';
 import { join } from 'path';
 
+import { remove, existsSync } from 'fs-extra';
+import { CommandCore } from '@midwayjs/command-core';
+import { BuildPlugin as ViteBuildPlugin } from '../src/cli/plugin/cli-plugin-build';
+import * as assert from 'assert';
+
+const cwd = join(__dirname, 'fixtures/base-react-app');
+const run = async (cwd: string, command: string, options = {}) => {
+  const core = new CommandCore({
+    commands: [command],
+    options: {
+      buildCache: true,
+      ...options,
+    },
+    log: {
+      log: console.log,
+    },
+    cwd,
+  });
+  core.addPlugin(ViteBuildPlugin);
+  await core.ready();
+  await core.invoke();
+};
+
+it('build ssr', async () => {
+  const dist = join(cwd, 'public/html');
+  if (existsSync(dist)) {
+    await remove(dist);
+  }
+  await run(cwd, 'build', {
+    viteConfigFile:'vite.config.ts',
+  });
+  assert(existsSync(join(dist, 'index.html')));
+})
+
 describe('/test/index.test.ts', () => {
   let app:Application;
 
@@ -30,6 +64,4 @@ describe('/test/index.test.ts', () => {
     expect(result.status).toEqual(200);
     expect(result.text).toContain('<h3>About</h3>')
   });
-
-  // it('production asset preloading', async () => {})
 })
